reflector reorganizado

+++ smart_drillholes_gui/mainapp/views.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.shortcuts import render, redirect
 from smart_drillholes.core import *
-from .forms import OpenForm, NewForm, AddTableForm
+from .forms import OpenSQliteForm, OpenPostgresForm, NewForm, AddTableForm
 import datetime
 
 
@@ -13,9 +13,10 @@ from sqlalchemy import Column, String, Float
 #from django.views.decorators.csrf import csrf_exempt
 #import json
 from reflector.og_reflector import Reflector
-from sqlalchemy.orm import sessionmaker
+
 import os
 import re
+from django.urls import reverse
 
 def index(request):
     response = render(request,
@@ -26,25 +27,49 @@ def index(request):
 
 def open(request):
     if request.method == "GET":
-        form = OpenForm()
-        return render(request,
-                      'mainapp/open.html',
-                      {'form': form,
-                       'ref': 'open'})
+        form = OpenSQliteForm()
+
     elif request.method == "POST":
-        form = OpenForm(request.POST)
+        if request.POST['db_type'] == "postgresql":
+            form = OpenPostgresForm(request.POST)
+        elif request.POST['db_type'] == "sqlite":
+            form = OpenSQliteForm(request.POST, request.FILES)
         if form.is_valid():
             if form.cleaned_data.get('db_type') == 'sqlite':
-                con_string = 'sqlite:///{}.sqlite'.format(form.cleaned_data.get('name'))
-            elif form.cleaned_data('db_type') == 'postgresql':
-                con_string = 'postgresql://postgres@localhost/{}'.format(form.cleaned_data.get('name'))
-            eng, meta = og_connect(con_string)
-            response = redirect('mainapp:dashboard')
+                #content_type: application/octet-stream
+                urlfile = request.FILES["sqlite_file"]
+                name = form.cleaned_data.get('sqlite_file')
+
+                BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+                dbName = BASE_DIR+'/smart4.sqlite'
+                if dbName != '':
+                    engineURL = 'sqlite:///'+dbName
+                #con_string = 'sqlite:///{0}.sqlite'.format(name)
+                con_string = engineURL
+            elif form.cleaned_data.get('db_type') == 'postgresql':
+                host = form.cleaned_data.get('host')
+                dbname = form.cleaned_data.get('name')
+                user = form.cleaned_data.get('user')
+                password = form.cleaned_data.get('password')
+                con_string = 'postgresql://{2}:{3}@{0}/{1}'.format(host,dbname,user,password)
+                #response = redirect('mainapp:dashboard')
+
+            request.session['engineURL'] = con_string
+
+            reflector = Reflector(con_string)
+            reflector.reflectTables()
+            cols,tks,data,table_key = update(reflector)
+
+            return render(request,'mainapp/reflector.html', {'tks': tks,'cols':cols,'data':data,'table_key':table_key})
+                #return render(request,'mainapp/open.html', {'urlfile':con_string})
             expiry_time = datetime.datetime.now() + datetime.timedelta(minutes=525600)
             response.set_cookie(key='db', value=form.cleaned_data.get('name'), expires=expiry_time)
             response.set_cookie(key='db_type', value=form.cleaned_data.get('db_type'), expires=expiry_time)
             return response
-
+    return render(request,
+                     'mainapp/open.html',
+                     {'form': form,
+                      'ref': 'open'})
 
 def new(request):
     if request.method == "GET":
@@ -59,7 +84,9 @@ def new(request):
             if form.cleaned_data.get('db_type') == 'sqlite':
                 con_string = 'sqlite:///{}.sqlite'.format(form.cleaned_data.get('name'))
             elif form.cleaned_data('db_type') == 'postgresql':
+
                 con_string = 'postgresql://postgres@localhost/{}'.format(form.cleaned_data.get('name'))
+
             eng, meta = og_connect(con_string)
             og_references(eng, meta, table_name='assay_certificate', key='SampleID', cols={'Au': {'coltypes': Float,
                                                                                            'nullable': True}})
@@ -76,8 +103,8 @@ def new(request):
                                                                                           'ondelete': 'RESTRICT',
                                                                                           'onupdate': 'CASCADE'}}})
 
-            #og_create_dhdef(eng,meta)
             execute(eng, meta)
+
             response = redirect('mainapp:dashboard')
             expiry_time = datetime.datetime.now() + datetime.timedelta(minutes=525600)
             response.set_cookie(key='db', value=form.cleaned_data.get('name'), expires=expiry_time)
@@ -93,6 +120,9 @@ def dashboard(request):
 
 #@csrf_exempt
 def reflector(request, table_key = ''):
+    engineURL = request.session.get('engineURL')
+    reflector = Reflector(engineURL)
+    reflector.reflectTables()
 
     if request.method == 'POST':
         pks = request.POST.getlist('checkbox-delete')
@@ -105,18 +135,13 @@ def reflector(request, table_key = ''):
 
         object_table = type(str(table_key), (Base,), defineObject(table))
 
-        for pk in pks:
-            query = session.query(object_table).get(pk)
-            session.delete(query)
-            session.commit()
-
-    BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    dbName = BASE_DIR+'/smart4.sqlite'
-    if dbName != '':
-        engineURL = 'sqlite:///'+dbName
-        global reflector
-        reflector = Reflector(str(engineURL))
-        reflector.reflectTables()
+        if pks:
+            session = reflector.make_session()
+            for pk in pks:
+                query = session.query(object_table).get(pk)
+                session.delete(query)
+                session.commit()
+            session.close()
 
     cols,tks,data,table_key = update(reflector, table_key)
 
@@ -147,14 +172,10 @@ def defineObject(table):
 
     return tbl_def
 
-def update(reflector, table_key):
+def update(reflector, table_key = '', session = ''):
     if reflector.is_reflected():
-        #table list
-        tableList = reflector.getOg_tables()
-        DBSession = sessionmaker(bind=reflector.get_engine())
-        #nonlocal session
-        global session
-        session = DBSession()
+        if session == '':
+            session = reflector.make_session()
 
         #table names for template
         tks = reflector.get_tableKeys()
@@ -180,7 +201,8 @@ def update(reflector, table_key):
 
             dic = {'pks':','.join(ids),'data':dt}
             data.append(dic)
-
+    session.close()
+    del reflector
     return (cols,tks,data,table_key)
 
 def add_table(request):
